+++ boot-data-redis/src/test/java/com/example/demo/ContainersConfig.java
@@ -1,0 +1,18 @@
+package com.example.demo;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.utility.DockerImageName;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class ContainersConfig {
+
+    @Bean
+    @ServiceConnection(name = "redis")
+    GenericContainer<?> redisContainer() {
+        return new GenericContainer<>(DockerImageName.parse("redis").withTag("7.0.11-alpine"))
+                .withExposedPorts(6379);
+    }
+}
