+++ boot-data-redis/src/test/java/com/example/demo/DemoApplicationTests.java
@@ -7,12 +7,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.webtestclient.autoconfigure.AutoConfigureWebTestClient;
 import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 @AutoConfigureWebTestClient
-@SpringBootTest
-@Import(TestDemoApplication.class)
+@SpringBootTest(classes = {DemoApplication.class, ContainersConfig.class})
 public class DemoApplicationTests {
 
   @Autowired
