+++ boot-data-redis/src/test/java/com/example/demo/TestDemoApplication.java
@@ -1,25 +1,12 @@
 package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.utility.DockerImageName;
 
-@TestConfiguration(proxyBeanMethods = false)
 public class TestDemoApplication {
-
-    @Bean
-    @ServiceConnection(name = "redis")
-    GenericContainer<?> redisContainer() {
-        return new GenericContainer<>(DockerImageName.parse("redis").withTag("7.0.11-alpine"))
-                .withExposedPorts(6379);
-    }
 
     public static void main(String[] args) {
         SpringApplication.from(DemoApplication::main)
-                .with(TestDemoApplication.class)
+                .with(ContainersConfig.class)
                 .run(args);
     }
